move the float matchers to a separate module for tidiness
